# api/v2/swagger/swagger.yaml
definitions:
  api.BusinessError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.Response:
    properties:
      data:
        type: object
      response:
        $ref: '#/definitions/api.BusinessError'
    type: object
...
info:
  contact:
    name: Singleton Team
    url: https://github.com/vmware/singleton
  description: This is a Singleton server
  license:
    name: EPL-2.0
    url: https://www.eclipse.org/legal/epl-2.0/
  title: Singleton Service API
  version: "2.0"
paths:
...
  /formatting/patterns:
    get:
      description: Get pattern data with language, region
      parameters:
      - description: a string which represents language, e.g. en,en-US,pt,pt-BR,zh-Hans
        in: query
        name: language
        required: true
        type: string
      - description: a string which represents region, e.g. US,PT,CN
        in: query
        name: region
        required: true
        type: string
      - description: pattern category string, separated by commas. e.g. 'dates,numbers,currencies,plurals,measurements,dateFields'
        in: query
        name: scope
        required: true
        type: string
      - description: a string for filtering the pattern data, separated by comma and
          underline. e.g. 'dates_eras,dates_dateTimeFormats'
        in: query
        name: scopeFilter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "206":
          description: Successful Partially
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get pattern data
      tags:
      - formatting-pattern-api
  /formatting/patterns/locales/{locale}:
    get:
      description: Get pattern data in a specified locale
      parameters:
      - description: locale String. e.g. 'en-US'
        in: path
        name: locale
        required: true
        type: string
      - description: pattern category string, separated by commas. e.g. 'dates,numbers,currencies,plurals,measurements,dateFields'
        in: query
        name: scope
        required: true
        type: string
      - description: a string for filtering the pattern data, separated by comma and
          underline. e.g. 'dates_eras,dates_dateTimeFormats'
        in: query
        name: scopeFilter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "206":
          description: Successful Partially
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get pattern data
      tags:
      - formatting-pattern-api
  /locale/regionList:
    get:
      description: Get region names in a specified locale
      parameters:
      - description: 'the supported language list, separated by commas. e.g. ''en,zh,ja'' '
        in: query
        name: supportedLanguageList
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "206":
          description: Successful Partially
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get region names
      tags:
      - locale-api
...
swagger: "2.0"
...